+++ src/components/automation/components/StopButton.jsx
@@ -1,13 +1,9 @@
-import React, { useContext } from 'react';
+import React from 'react';
 import { IoStop } from 'react-icons/io5';
 import { motion } from 'framer-motion';
-import { logEvent, stopIdler } from '@/src/utils/utils';
-import { toast } from 'react-toastify';
-import { AppContext } from '../../layout/components/AppContext';
+import { stopIdler } from '@/src/utils/utils';
 
-export default function StopButton({ isMountedRef, abortControllerRef, gamesWithDrops, screen, currentGame }) {
-    const { setActivePage } = useContext(AppContext);
-
+export default function StopButton({ setActivePage, isMountedRef, abortControllerRef, gamesWithDrops, screen, currentGame }) {
     const borderWidths = [
         ...Array.from({ length: 500 }, (_, i) => 0.5 + i * 0.01),
         ...Array.from({ length: 500 }, (_, i) => 6 - i * 0.01),
@@ -23,9 +19,7 @@ export default function StopButton({ isMountedRef, abortControllerRef, gamesWith
                 await stopIdler(currentGame.appId, currentGame.name);
             }
         } catch (error) {
-            toast.error(`Error in (handleStop): ${error?.message || error}`);
             console.error('Error in (handleStop) :', error);
-            logEvent(`[Error] in (handleStop) ${error}`);
         } finally {
             isMountedRef.current = false;
             abortControllerRef.current.abort();