+++ src/components/gameslist/components/AutomateButtons.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Button } from '@nextui-org/react';
-import { TbCardsFilled } from "react-icons/tb";
+import { TbCardsFilled } from 'react-icons/tb';
 import { FaAward } from 'react-icons/fa';
 import { useAutomate } from '../hooks/useAutomateButtons';