+++ changelog.md
@@ -1,4 +1,7 @@
 [![Downloads](https://img.shields.io/github/downloads/zevnda/steam-game-idler/1.6.3/total?style=for-the-badge&logo=github&color=137eb5)](https://github.com/zevnda/steam-game-idler/releases/download/1.6.3/Steam.Game.Idler_1.6.3_x64_en-US.msi)
 
 ## Changelog
-- Fixed an issue causing the timer on the `card farming` screen to be set multiple times when farming more than one game
+- Fixed an issue causing the `notifications` panel to be hidden by the `games list` page header
+- Reverted changes made in `1.6.0` to the `card farming` and `achievement unlocker` components that were causing the timer to get stuck at `00:01/00:00`
+  - #54 by @ialsca
+  - #58 by @Jesewe